Initial commit.
This commit includes all files contained in the latest bundle fetched from
shinyapps.io for Tool1v3En. See file metadata.yml for more information.

The purpose is to start a new repository from the latest version currently
in production. Files will soon change, and the project will be restructured,
and refactored. It will always be possible to revert this (initial) version.