New as_percentage() function.
This is a helper function that converts numeric values to character
vectors (percentages with 3 significant digits). This functions will
replace all these weird formatting mechanisms found in the server
component of the app.

Here is an example: return(paste(signif(res4,3)," %",sep="")). While
this works, as_percentage() is more digestable in the context of app.R
and is also much faster (it relies on sprintf()).