Remove old i18n mechanisms (WIP).
Attempting to launch the application as it currently is in this commit
won't work. This is because calls to deprecated functions gett() and
gettt() were naively replaced by calls to transltr::translate(). The
latter requires a Translator object, which will be introduced in a
future commit.

For now, I am focusing on integrating the source text back into the
source code.