Reintroduce source text in app and refactor.
This is a really huge commit introducing a very large set of major
changes. The whole source code changed, and it would be impossible
to list in details what was done.

The objective was to introduce back into the source code the actual
source text. Older translation mechanisms used labels to fetch text
to translate from external files. Consequently, the whole project
became a mess over time. It became really hard to maintain the
application and I spent quite a long time analyzing, understanding,
testing, documentating, and reorganizing code. Along the way, I
chose to fix many bugs, typos, errors, inconsistencies, etc.

The end result is a refactored app that works like the previous
version. Its structure and its UI is mostly unchanged from the
user's perspective. I introduced quite a lot of small changes,
but nothing fancy. The app performs the same thing as previously
and shows results in the same way.

Under the hood, version 4 is a major revision. Almost everything
changed for the better. I made sure no existing functionalities
were broken following the changes (manually).

The app (mostly the ui object) will change again in a near
future. Calls to translate() will move to server(). This is
the last step for the internationalization process.

Some custom functions in R still require documentation.

There is a new Markdown document for inputId and outputId values.
The goal is to bring order and consistency to improve readability.
The actual changes (to v4) must still be done.