PR #2: Partial refactoring of the project.
## Changes

This PR introduces a large set of major changes. While the application
itself did not change *much* from the users' perspective, it drastically
changed *under the hood*. As explained, the application itself got a
major overhaul that was necessary (1) to ease its long-term maintenance
and (2) to introduce `transltr` mechanisms and support multiple
languages.

The refactoring efforts focused on the file `app.R` (which combines
older scripts `ui.R` and `server.R` mainly). Other files were for the
most part left as is (unless minor changes were absolutely necessary).
Multiple new files were introduced in the process.

Attempting to exhaustively list all changes is not possible. (there were
that many changes). Below is a summary.

- The UI got a little overhaul with a more consistent set of CSS
classes.
- Old translations / internationalization mechanisms were removed.
- Unused files were removed.
- Google Analytics was deactivated until further notice. See the section
in the documentation of the `app.R` script.
- New script to load global constants and dependencies (scripts and
libraries).
- A new set of functions to create the UI.
- Replaced all previous CSS classes with modern alternatives that are
(more) semantic, simpler to use and simpler to maintain.
- Removed all JS logic and external files. The logic was reimplemented
in R.
- Changed all identifiers (`inputId` and `outputId`) for consistency.
See new file `IOmd` for more information.
- Completely rewrote the `ui` object.
- Rewrote many parts of `server()`. However, what it accomplishes (and
functions it calls defined in `.scripts/`) were left as is.
- `ui` and `server()` were reorganized accordingly.

## `FIXME`

I left multiple `FIXME` notes in `app.R`. These are tasks that should be
done but that are outside of my (current) scope / quote. They currently
have no immediate impact on the app (thy do not target bugs).

## Notes

The application is now stable and works as intended, both locally and
remotely. The next PR will introduce support of multiple languages.

Users having access can already see the new version in production:
https://lavoue.shinyapps.io/tool1/.