Remove input al (exposure limit multiplier).
The objective was to stop using internal 'c.oel' in favor of 'oel'
(input oel as is). Precedently, c.oel was used for inference and it
was equal to oel * oel.mult (the latter was set equal to input$al).

To avoid modifying common scripts stored in scripts/, the value
passed to argument oel.mult of data.formatting.SEG() is internally
set equal to 1. Therefore, c.oel always equals oel. All other
functions that used c.oel via reactive value user_formatted_sample()
now directly uses input$oel.

This is not perfect because we can still see c.oel in the source
code (and this can be confusing), but it is better than nothing.
At the very least, users don't have access to input$al now, and
Webexpo scripts are left as is.