New submit button to manually start calculations.
This is a first iteration on a usual submit button to let users
control when bayesian calculations should be done (instead of
being done automatically whenever any input except lang changes).

It is implemented via a straightforward Shiny button. The core
reactive value user_formatted_sample() is now updated only when
this button is clicked. One downside of this approach is that
static parts of the UI is shown anyway, which is not optimal. I
am currently investigating solutions to this.

- Add tooltips for submit and clear buttons.
- Change top banner message.
- Update subsection How to Use of panel About following these
  changes.