Use a global persistent cache for truly static UI objects.
UI elements that only depend on the current language are now
cached and reused (instead of being re-rendered over and over
again). This is accomplished by using shiny::bindCache() with
the output of lang() being the key.

The cache is shared by all sessions, never expires, and is
never deleted (until a new version is deployed). Therefore,
they all contribute to it.

It is hard to assess with confidence whether this is truly
beneficial or not. Early tests shows that it may improve
performance (slightly), espacially when the Shiny instance
has to deal with multiple concurrent sessions/users. If we
later see that bindCache() causes more problems than it
attempts to solve, it will be deactivated.