New FAQ Modal panel.
It replaces the older About Modal panel. It also replaces the old
Information card previously embedded in the Descriptive Statistics
panel.

The objective of this new panel is to have a central location for
all current and future documentation of Tool 1. It is designed to
scale in time and accomodate all topics (in a much easier way).

Details are thoroughly documented. See the underlying source script.