Simplify implementation of ui_title().
This is undocumented, but bslib::page_sidebar() automatically
wraps the content of what is passed to argument `title` in a
div.navbar.navbar-static-top > div.container-fluid container:

<div class="navbar navbar-static-top">
  <div class="container-fluid">
    ... [content of `title` goes here in output]
  </div>
</div>.

Therefore, the code can be simplified by removing unnecessary
containers and return a list of `shiny.tag` objects instead.
These elements can be styled with Bootstrap navbar-* classes
directly. This is a first step towards creating a mobile-friendly
version of Tool 1.