Fix: ensure that cards in Sidebar module grow as expected.
They no longer collapse to a single line with a width equal to 2px
(top and bottom borders). This used to happen because cards are
both fill items and fillable containers. Moreover, bslib::sidebar()
objects are fillable by design, so this was causing unexpected
behavior. Encapsulating bslib::cards() in a single <div> solves the
problem.