PR#6: Introduce responsive elements to support a wide variety of screens.
## Release

This PR is tied to the `4.1.0` **minor** release (`2025-04-16`).

## Changes

This PR introduces minor changes to support a wide variety of screens,
including small cellphones. A device screen having a width of at least
`375` pixels is recommended, but the application should be able to
support even smaller screens.

Most elements were already responsive. Most of the work focused on the
Title Bar module (which is fully custom).

This PR further introduces small optimizations for faster page loading
when `lang` is `"en"`.

Other changes are listed in a new `NEWS.md` file.