Bugfix: do not use Shiny built-in auto-reload feature.
- It is partially broken (this is documented in opened PRs of the
  source repository of shiny, as of April/May 2025).
- It resets connections to the app locally when enabled and when
  the active panel is Exceedance Fraction.

This is likely attributable to the fact it detects changes in
files stored in www/images and reloads the app accordingly.
However, said changes are not 'true' changes. Files do get
replaced (at least in the current implementation), but by
identical ones yielding no effective change from the perspective
of the user.

The auto-reload feature of Shiny seems convenient on paper.
In practice, it is buggy, a pain in the ass to manage, and
broken.

Removed and disabled forever.