Simplify implementation of  buttons and observers for languages.
The process is now divided in three manual steps instead of being
fully automated. While it may not appear optimal to some developers,
doing so dramatically simplifies the logic (and future maintenance).
See the underlying documentation for more information.

The process should take less than 5 minutes (which is fast enough
to justify a manual process). The only requirement is be careful
and ensure identifiers and labels match what tr$native_languages
returns.