Bugfix: find-text.R script was not dropping older translations.
The logic of this script had to be rewritten. The previous
implementation was not dropping older (unused) pairs of source
text and translation. Instead, translation files were accumulating
them. This was problematic because whenever the source text had to
be modified, olver versions would be kept. Translation files became
bloated and messy.

The new version creates a fresh Translator object from the source
code and then import back translations that are still relevent.
Those that are not are discarded when translator_write() is called
(files are overwritten).