Bugfix: deprecate ui_element(). Modify ui_link_*().
Usage of placeholders in the source text should be minimized. It
may be otherwise be really hard to translate and may impact how
much time is required to translate the app.

In order to simplify the translation process (and the code itself),
ui_element() is removed.

I also added a new ui_bs_color() function to validate Bootstrap
color names. This function is called by ui_link() functions.