Reintegrate the OEL multiplier input.
This input was previously named 'al' in earlier versions and was
removed on February 26, 2025 in commit #fc91cf3, as requested by
@jlavoue:

https://github.com/webexpo/tool1/commit/fc91cf3e5dc5732f77572cba382fc0dcfde27b0e.

Following recent discussions with users, it was decided that this
input would be reintegrated into version 5 of Tool 1 (and released
as part of the upcoming version 5.1).

This commit is for the most part a manual revert of #fc91cf3, with
further modifications to comply with the actual state of the code.

The input was renamed oel_multiplier in the process and can only be
used in the extended mode.