Update how static assets are managed.
The objective of this commit is to ease the management of relative
file paths with rmarkdown. File paths are usually relative to the
underlying source Markdown files being converted with
rmarkdown::render(). This makes it hard to reuse fragments
with files located in different subdirectories of the project.

However, fragments are required to include tags in the <head> of
self contained documents generated with rmarkdown::render(). This
is because a YAML header cannot be used in the source documents,
because their rendered versions on GitHub is super ugly.

Hence, the changes below were required.

1. All Markdown files requiring a static HTML equivalent file
   are now stored at top-level.
2. File intl/README.md was moved to TRANSLATIONS.md because of (1).
3. Both NEWS.md and TRANSLATIONS.md now use the same styles
   (www/static/_main.css) and the same <head> template
   (www/assets/_head.html). These are required for consistency with
   Tool 1. The latter is required to include favicons into self-
   contained documents.
4. Subdirectory www/static was renamed to www/assets. This is because
   www/ only contains static assets. However, www/assets contains
   static documents other than usual web files (favicons, styles,
   JS scripts, and web manifest).
5. All other files in the source code were updated accordingly.