New experimental mechanism for translatable URLs [WIP].
url() is designed to mark character strings as such. By default,
the URL passed to it is assigned to the current default language
(typically "en"). Further character strings can be passed to ...
These are alternative URLs pointing to the same entity (but in
other languages).

The design is experimental and may change.