New mechanisms to store and fetch AIHA risk levels.
Risk levels are now encoded as integers in the code. Information
on each one can be fetched via functions get_risk_level_*(). This
allows a more robust encapsulation of internal constants used
throughout the code base.

There is one exception allowed. The risk level names (acceptable,
tolerable, and problematic) can be used as is in the source text.
This is for convenience.