PR #12: Integrate logic for Spanish translations and many backend changes.
## Release

This PR is tied to the next minor release ( `5.2.0` as of `2025-07-18`).
It introduces many changes on the backend side and the required logic
for Spanish translations. Since they are still being reviewed, they are
not yet visible and will be integrated in the
next version (likely `5.2.1`).

## Changes

Changes are listed and described succintly in file
[NEWS.md](https://github.com/webexpo/tool1/blob/2a5d04147c0888a39d244d44cf26a40eed101d73/NEWS.md).