Bugfix: allow translation of categories in fun.qqplot.group.D().
Previously, this function did not allow translation of the lengend's
title for categories ("Category"). This is now fixed thanks to new
argument label_category and underlying logic with
ggplot2::guide_legend().

While this function is not used by tool1, I'm still adding the fix
here for consistency between tool1 and tool3. This is better until
the official Expostats scripts get their own repository.